[Schema transformer] Fix error case when object disappear
During SNAT tests where I set and clear router's gateway interface in a
loop (that creates and deletes a virtual network, a service instance, virtual
machines with VMIs, a network policy and a routing table each time), I
get some issues with the schema transformer which tries to access
resources that already gone.
That patch regroups the fixes I should done to not crash the schema
transformer during that tests.

Change-Id: Iebed35eb64e2541e317230b628812db97bd55020
Signed-off-by: Édouard Thuleau <edouard.thuleau@cloudwatt.com>